Prevent backflush when sanitizing inputs.
This prevents an error condition in the Glimmer2 rendering engine.
Basically, since `value` has already flowed downwards causing it to be
set (and trigger a re-render) during the same render flow is an error.

This avoids the error condition by adding a small amount of async before
invoking the `update` function with the sanitized value.