action "update" called in own method
Hi,
I'd like to move the execution of `invokeAction(this, 'update', value)` from `_processNewValue` to its own method.

It doesn't break the behaviour of the inputs as it is not impactful on the flow of `_processNewValue`.
An advantage is that it allows to subclass the components and add new behaviours such as debouncing (my current use-case for this PR).